+++ cli/src/main.rs
@@ -88,8 +88,13 @@ struct Tweaks {
     /// The number of backtracking stages. Backtracking prevents 'garbage' generation.
     #[structopt(long = "backtrack-stages", default_value = "5")]
     backtrack_stages: u32,
-    /// Show a window with the current progress of the generation
     #[structopt(long = "window")]
+    #[cfg(feature = "progress")]
+    #[cfg_attr(feature = "progress", structopt(long = "window"))]
+    #[cfg_attr(
+        feature = "progress",
+        doc = "Show a window with the current progress of the generation"
+    )]
     show_window: bool,
     /// Show a window with the current progress of the generation
     #[structopt(long)]
@@ -264,10 +269,14 @@ fn real_main() -> Result<(), Error> {
     let progress: Option<Box<dyn texture_synthesis::GeneratorProgress>> =
         if !args.tweaks.no_progress {
             let progress = ProgressWindow::new();
-            let progress = if args.tweaks.show_window {
-                progress.with_preview(args.out_size, std::time::Duration::from_millis(100))?
-            } else {
-                progress
+
+            #[cfg(feature = "progress")]
+            let progress = {
+                if args.tweaks.show_window {
+                    progress.with_preview(args.out_size, std::time::Duration::from_millis(100))?
+                } else {
+                    progress
+                }
             };
 
             Some(Box::new(progress))
@@ -293,13 +302,12 @@ fn real_main() -> Result<(), Error> {
 }
 
 use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
+#[cfg(feature = "progress")]
 use minifb::Window;
-use std::time::Duration;
 
 pub struct ProgressWindow {
-    window: Option<Window>,
-    update_freq: Duration,
-    last_update: std::time::Instant,
+    #[cfg(feature = "progress")]
+    window: Option<(Window, std::time::Duration, std::time::Instant)>,
 
     total_pb: ProgressBar,
     stage_pb: ProgressBar,
@@ -330,9 +338,8 @@ impl ProgressWindow {
         });
 
         Self {
+            #[cfg(feature = "progress")]
             window: None,
-            update_freq: Duration::from_millis(10),
-            last_update: std::time::Instant::now(),
             total_pb,
             stage_pb,
             total_len: 100,
@@ -341,16 +348,21 @@ impl ProgressWindow {
         }
     }
 
-    fn with_preview(mut self, size: (u32, u32), update_every: Duration) -> Result<Self, Error> {
+    #[cfg(feature = "progress")]
+    fn with_preview(
+        mut self,
+        size: (u32, u32),
+        update_every: std::time::Duration,
+    ) -> Result<Self, Error> {
         let window = Window::new(
             "Texture Synthesis",
             size.0 as usize,
             size.1 as usize,
             minifb::WindowOptions::default(),
-        )?;
+        )
+        .unwrap();
 
-        self.window = Some(window);
-        self.update_freq = update_every;
+        self.window = Some((window, update_every, std::time::Instant::now()));
 
         Ok(self)
     }
@@ -380,31 +392,36 @@ impl texture_synthesis::GeneratorProgress for ProgressWindow {
         self.total_pb.set_position(update.total.current as u64);
         self.stage_pb.set_position(update.stage.current as u64);
 
-        if let Some(ref mut window) = self.window {
-            let now = std::time::Instant::now();
+        #[cfg(feature = "progress")]
+        {
+            if let Some((ref mut window, ref dur, ref mut last_update)) = self.window {
+                let now = std::time::Instant::now();
 
-            if now - self.last_update < self.update_freq {
-                return;
+                if now - *last_update < *dur {
+                    return;
+                }
+
+                *last_update = now;
+
+                if !window.is_open() {
+                    return;
+                }
+
+                let pixels = &update.image;
+                if pixels.len() % 4 != 0 {
+                    return;
+                }
+
+                // I guess someone using ARM might not like this part
+                let pixels: &[u32] = unsafe {
+                    let raw_pixels: &[u8] = pixels;
+                    #[allow(clippy::cast_ptr_alignment)]
+                    std::mem::transmute(&*(raw_pixels as *const [u8] as *const [u32]))
+                };
+
+                // We don't particularly care if this fails
+                let _ = window.update_with_buffer(pixels);
             }
-
-            if !window.is_open() {
-                return;
-            }
-
-            let pixels = &update.image;
-            if pixels.len() % 4 != 0 {
-                return;
-            }
-
-            // I guess someone using ARM might not like this part
-            let pixels: &[u32] = unsafe {
-                let raw_pixels: &[u8] = pixels;
-                #[allow(clippy::cast_ptr_alignment)]
-                std::mem::transmute(&*(raw_pixels as *const [u8] as *const [u32]))
-            };
-
-            // We don't particularly care if this fails
-            let _ = window.update_with_buffer(pixels);
         }
     }
 }