+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - CLI: Replaced piston_window with [`minifb`](https://crates.io/crates/minifb)
+- CLI: Due to how minifb works via X11, the progress window is now an optional feature
+not enabled when building for musl
 
 ### Removed
 - Removed several codec features from `image`, only `png`, `jpeg`, `bmp`, and `tga` are supported now