+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] - 2019-09-04
 ### Changed
 - Use [`failure`](https://crates.io/crates/failure) for errors
 - CLI: Replaced piston_window with [`minifb`](https://crates.io/crates/minifb)
@@ -55,7 +57,8 @@ callers
 ### Added
 - Initial add of `texture-synthesis`
 
-[Unreleased]: https://github.com/EmbarkStudios/texture-synthesis/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/EmbarkStudios/texture-synthesis/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/EmbarkStudios/texture-synthesis/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/EmbarkStudios/texture-synthesis/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/EmbarkStudios/texture-synthesis/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/EmbarkStudios/texture-synthesis/releases/tag/0.1.0